Added Move Methods
Using two different ways to implement the say.  Any point on a screen can be given with an x and or a y coordinate.  These can be passed quickly as an integer.  Howver in many cases a Point object is being tracked and can be easily passed to the move function as well.