Defensive Programming
Here we look for a null value to be passed and throw an Argument Null Exception.  Let's run the code now.  {CTRL} + {F5} the code runs and gets a different error this time.  Now an error indicating that an Argument Null Exception was thrown and the Parameter name is newLocation.